+++ atlite/resource.py
@@ -67,7 +67,7 @@ def get_windturbineconfig(turbine, add_cutout_windspeed=False):
     if add_cutout_windspeed is False:
         msg = (
             "'add_cutout_windspeed' for wind turbine\npower curves will default to "
-            "True in atlite relase v0.2.13."
+            "True in atlite relase v0.2.15."
         )
         warnings.warn(msg, FutureWarning)