+++ RELEASE_NOTES.rst
@@ -11,6 +11,10 @@ Release Notes
 Upcoming Release
 ================
 
+- Adds option to toggle whether ERA5 downloads are requested in monthly or
+  annual chunks with keyword argument ``cutout.prepare(monthly_requests=True)``.
+  The default is now annual requests.
+
 - Improved parallelization of ``atlite.convert.build_line_rating`` by adding
   keyword arguments for ``dask.compute`` (``dask_kwargs={}``) and an option to
   disable the progressbar (``show_progressbar=True``).